update grafana dashboard & create corbench image
