+++ corbench/Dockerfile
@@ -1,17 +1,15 @@
-FROM prominfra/infra:master
+FROM --platform=linux/amd64 golang:1.15-alpine
 
-LABEL maintainer="The Prometheus Authors <prometheus-developers@googlegroups.com>"
+WORKDIR /corbench
 
-WORKDIR /prombench
-
-ENV INFRA_CMD infra
+RUN apk add git make
 
 # Copy Makefiles and manifests
 # Need 'cd' since ghActions ignores WORKDIR
 # Need 'eval' to prevent bash keywords be run as commands
 COPY ./ ./
 
-RUN echo -e '#!/bin/sh\ncd /prombench\neval "$@"' >/bin/docker_entrypoint
+RUN echo -e '#!/bin/sh\ncd /corbench\neval "$@"' >/bin/docker_entrypoint
 
 RUN chmod u+x /bin/docker_entrypoint
